added error processing. 'id' parameter changed to 'telegram id'

+++ src/routes/delteteUser.ts
@@ -1,27 +1,31 @@
 import express from 'express';
-import {deleteUser} from "../bd";
 import {User} from "../models";
 
 const router = express.Router();
 
-router.delete('/:id', (req, res) => {
-    User.destroy({
-        where: {
-            id: req.params.id
+router.delete('/', async (req, res) => {
+    try{
+        if (req.body.tg_id === undefined) {
+            throw new Error('Telegram id (tg_id) parameter is not found');
         }
-    })
-        .then(data => {
-            res.status(200).json({
-                status: 'success',
-                info: `user ${req.params.id} has been deleted`
-            })
+        const isDeleted = await User.destroy({
+            where: {
+                tg_id: req.body.tg_id
+            }
         })
-        .catch(error => {
-            res.status(404).json({
-                status: 'error',
-                error
-            })
+        if(!isDeleted){
+            throw new Error('User Does not exist')
+        }
+        res.status(200).json({
+            status: 'success',
+            info: `user ${req.body.tg_id} has been deleted`
         })
+    } catch (error) {
+        res.status(404).json({
+            message: error.message,
+            error: error
+        })
+    }
 })
 
 module.exports = router;
